`w` was already `usize`, why cast

+++ src/draw.rs
@@ -36,7 +36,7 @@ pub fn point_at_arg(mut idx: usize, style: &Style, specific_idx: Option<usize>) 
     let (spacing, arg_len) = SPACES[&idx];
 
     if let Some(w) = *TERM_WIDTH
-        && JOINED_ARGS.len() > w as usize
+        && JOINED_ARGS.len() > w
     {
         let print_width = (w - 20).min(arg_len);
         let arg: &str = &RAW_ARGS[idx][0..print_width];
